# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elepad API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "cause": {
                "nullable": true
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "GameListItem": {
        "type": "object",
        "properties": {
          "gameType": {
            "$ref": "#/components/schemas/GameType"
          },
          "gameName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "comingSoon": {
            "type": "boolean"
          }
        },
        "required": [
          "gameType",
          "gameName",
          "displayName",
          "description",
          "icon",
          "isAvailable"
        ]
      },
      "GameType": {
        "type": "string",
        "enum": [
          "memory",
          "logic",
          "attention",
          "reaction"
        ]
      },
      "PuzzleCreatedResponse": {
        "type": "object",
        "properties": {
          "puzzle": {
            "$ref": "#/components/schemas/Puzzle"
          }
        },
        "required": [
          "puzzle"
        ]
      },
      "Puzzle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gameType": {
            "$ref": "#/components/schemas/GameType"
          },
          "gameName": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "integer",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "gameType",
          "gameName",
          "title",
          "difficulty",
          "createdAt"
        ]
      },
      "NewMemoryPuzzle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "difficulty": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "rows": {
            "type": "integer",
            "minimum": 2,
            "maximum": 6,
            "default": 4
          },
          "cols": {
            "type": "integer",
            "minimum": 2,
            "maximum": 6,
            "default": 6
          }
        }
      },
      "NewNetPuzzle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "difficulty": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "gridSize": {
            "type": "integer",
            "minimum": 3,
            "maximum": 7,
            "default": 5
          }
        }
      },
      "NewFocusPuzzle": {
        "type": "object",
        "properties": {
          "rounds": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 10
          }
        }
      },
      "NewSudokuPuzzle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ],
            "default": "easy"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/games": {
      "get": {
        "tags": [
          "games"
        ],
        "responses": {
          "200": {
            "description": "Lista de juegos disponibles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameListItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/games/{gameName}": {
      "get": {
        "tags": [
          "games"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameName",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del juego",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListItem"
                }
              }
            }
          },
          "404": {
            "description": "Juego no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/puzzles/memory": {
      "post": {
        "tags": [
          "puzzles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMemoryPuzzle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Puzzle de memoria creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/puzzles/net": {
      "post": {
        "tags": [
          "puzzles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewNetPuzzle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Puzzle de NET creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/puzzles/focus": {
      "post": {
        "tags": [
          "puzzles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFocusPuzzle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Puzzle de atención creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/puzzles/sudoku": {
      "post": {
        "tags": [
          "puzzles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSudokuPuzzle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Puzzle de Sudoku creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/puzzles/recent/{gameType}": {
      "get": {
        "tags": [
          "puzzles"
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/GameType"
            },
            "required": true,
            "name": "gameType",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de puzzles recientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Puzzle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}